UNTESTED REVISION ==============================
First commit after BenH_FR:s observed that I didn't conform to the JSON-RPC spec. This will hopefully bring us closer without losing that much backwards-compatability.

This has not been tested yet, so it's strictly for those wanting to test the new features/bugfixes.

What's changed
-----------------------

POST requests made using the content type 'application/json' will be treated according the spec. Other POST requests will be treated as before. That is that the request will be assumed to be 'application/x-www-form-urlencoded' and the params parameter will be json-unserialized and the rest of the parameters will be passed to JsonRpcServer unchanged.

Breaking changes
-----------------------

The only breaking changes should be the GET-request handling. The get request was previously treated exactly as the POST-request. But the 1.1 spec states the following for JSON-RPC GET-requests: "The client MUST NOT use the JSON syntax for the parameter values". So that's been changed. Get requests must now be made as follows 'http://yoursite.com/services/json-rpc/method_name?param=val&param2=val". What's missing here is that identically named parameters aren't collapsed into arrays as per the spec. That will have to be fixed at a later date.